Refractored create-kaniko-job and modularised Artifacts secret creation process
